Y2038: correctly display neighbour in cachemgr
Ensure that the timeout value in the neighbours
cachemgr report is not truncated on systems
having 32-bit integers

+++ src/neighbors.cc
@@ -1453,7 +1453,7 @@ dump_peer_options(StoreEntry * sentry, CachePeer * p)
         storeAppendPrintf(sentry, " ttl=%d", p->mcast.ttl);
 
     if (p->connect_timeout_raw > 0)
-        storeAppendPrintf(sentry, " connect-timeout=%d", (int)p->connect_timeout_raw);
+        storeAppendPrintf(sentry, " connect-timeout=%ld", p->connect_timeout_raw);
 
     if (p->connect_fail_limit != PEER_TCP_MAGIC_COUNT)
         storeAppendPrintf(sentry, " connect-fail-limit=%d", p->connect_fail_limit);
